scripts/dtc: Export YYLOC global declaration
Building with GCC 10 gives this error:

 /usr/bin/ld: scripts/dtc/dtc-parser.tab.o:(.bss+0x50): \
   multiple definition of `yylloc'; scripts/dtc/dtc-lexer.lex.o:(.bss+0x0): first defined here
   collect2: error: ld returned 1 exit status
   make[4]: *** [scripts/Makefile.host:100: scripts/dtc/dtc] Error 1

commit e33a814e772c ("scripts/dtc: Remove redundant YYLOC global
declaration") fixes this in Mainline by removing the symbol
altogether.  However, this causes other 'not defined' issues in
android-3.18.

The original commit does mention that "the proper solution would be to
to mark this [symbol] as extern", which does not work in Mainline, but
does seem to satisfy android-3.18.

Signed-off-by: Lee Jones <lee.jones@linaro.org>
Change-Id: I4b5e7c8ae477c4eb6470472aad51ebddb128d7be